feat(vim): replace NERDTree with nvim-tree and add new plugins
This commit replaces the NERDTree plugin with nvim-tree for better
performance and integration with Neovim. It also adds new plugins such
as indent-blankline and nvim-treesitter for improved code readability
and syntax highlighting. The key mappings have been updated to reflect
these changes. The init script has been updated to remove an unnecessary
command related to Packer.